Added Cmake stuff for building the unit tests within trilinos To incorporate all of this within Trilinos what is left is to rename everthing "PB_" to "Teko" and add it to the Trilinos cvs repository. I will probably do the renaming using a branch on the local cvs repository and renaming in a safe testable way.

+++ packages/teko/src/PB_BlockLowerTriInverseOp.cpp
@@ -15,7 +15,7 @@ BlockLowerTriInverseOp::BlockLowerTriInverseOp(BlockedLinearOp & L,const std::ve
    int blocks = blockRowCount(L_);
    TEUCHOS_ASSERT(blocks>0);
    TEUCHOS_ASSERT(blocks==blockColCount(L_));
-   TEUCHOS_ASSERT(blocks==invDiag_.size());
+   TEUCHOS_ASSERT(blocks==(int) invDiag_.size());
 
    // create the range and product space
    ///////////////////////////////////////////////////
